+++ lib/model/news_data.dart
