+++ lib/provider/terminology_quiz_provider.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'dart:math';
+import '../service/user_service.dart';
 
 class TerminologyQuizService with ChangeNotifier {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+  final UserService _userService = UserService();
   int _currentQuestionIndex = 0;
   int _score = 0;
   bool _answered = false;
@@ -107,6 +109,10 @@ class TerminologyQuizService with ChangeNotifier {
         'completedAt': Timestamp.now(),
         'completed': newCompleted,
       });
+
+      if (newCompleted && !wasPreviouslyCompleted) {
+        await _userService.updateUserBalance(_uid, 100000); // Add 100,000 reward if completed
+      }
     } catch (e) {
       print('Error saving quiz completion: $e');
     }
