+++ lib/service/profile_service.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import '../model/user_model.dart';
+import '../model/user_data.dart';
 
 class ProfileService {
   final FirebaseAuth _auth = FirebaseAuth.instance;
