+++ lib/view/article/article_list_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:motu/view/article/widget/article_list_builder.dart';
 import '../../model/article_data.dart';
-import 'article_detail.dart';
 
 class ArticleListScreen extends StatelessWidget {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
@@ -96,7 +96,7 @@ class ArticleListScreen extends StatelessWidget {
                   delegate: SliverChildBuilderDelegate(
                         (context, index) {
                       final article = articles[index];
-                      return NewsCard(article: article);
+                      return articleListBuilder(context, article); // 변경된 부분
                     },
                     childCount: articles.length,
                   ),
@@ -105,86 +105,6 @@ class ArticleListScreen extends StatelessWidget {
             );
           }
         },
-      ),
-    );
-  }
-}
-
-class NewsCard extends StatelessWidget {
-  final Article article;
-
-  NewsCard({required this.article});
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => ArticleDetailScreen(article: article),
-          ),
-        );
-      },
-      child: Card(
-        margin: EdgeInsets.all(8.0),
-        child: Row(
-          children: [
-            // 이미지
-            Container(
-              width: 90,
-              height: 90,
-              padding: EdgeInsets.all(8.0),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(8.0),
-                child: Image.network(
-                  article.imageUrl,
-                  fit: BoxFit.cover,
-                ),
-              ),
-            ),
-            Expanded(
-              child: Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Container(
-                      margin: EdgeInsets.only(bottom: 8.0),
-                      child: Wrap(
-                        spacing: 8.0,
-                        children: article.topics.map((topic) {
-                          return Container(
-                            padding: EdgeInsets.symmetric(horizontal: 8, vertical: 2),
-                            decoration: BoxDecoration(
-                              color: Color(0xff701FFF),
-                              borderRadius: BorderRadius.circular(12),
-                            ),
-                            child: Text(
-                              topic,
-                              style: TextStyle(
-                                color: Colors.white,
-                                fontSize: 10,
-                              ),
-                            ),
-                          );
-                        }).toList(),
-                      ),
-                    ),
-                    // 제목
-                    Text(
-                      article.title,
-                      style: TextStyle(
-                        fontWeight: FontWeight.bold,
-                        fontSize: 16,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
       ),
     );
   }
