+++ lib/view/scenario/scenario_list.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:motu/service/scenario_service.dart';
+import 'package:motu/provider/scenario_service.dart';
 import 'package:motu/view/scenario/scenario_chart.dart';
 
 class ScenarioList extends StatelessWidget {
