+++ lib/view/terminology/widget/bookmark_card_builder.dart
