+++ lib/view/terminology/widget/terminology_card_builder.dart
