+++ lib/widget/bottom_nav_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:motu/service/navigation_service.dart';
+import 'package:motu/provider/navigation_provider.dart';
 import 'package:provider/provider.dart';
 import '../view/theme/color_theme.dart';
